backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
openai
requests
pathway
pandas
csv-diff
slowapi
python-jose[cryptography]
pymongo
certifi
bcrypt
email-validator
